requirements.txt
chainlit
literalai
SpeechRecognition
elevenlabs
PyAudio
pydub
numpy
openai
google-genai
python-dotenv
langchain
langsmith
langgraph
langchain-core
langchain-community
langchain-tavily
langchain-chroma
langchain-openai
langchain-ollama
langchain-google-genai
PyPDF2
python-docx
pillow
scikit-learn
beautifulsoup4
html2text
duckduckgo-search
ddgs
librosa
scipy
noisereduce
asyncpg
SQLAlchemy
aiohttp
gTTS
pandas
azure-identity 
azure-storage-file-datalake 
greenlet
azure-ai-inference
weasyprint
markdown
markdownify